development/requirements.txt
absl-py==0.2.0
aniso8601==3.0.0
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Cython==0.28.2
Flask==1.0.2
flask-restplus==0.10.1
gunicorn==19.8.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Keras==2.1.6
keras-resnet==0.1.0
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
mkl-fft==1.0.0
mkl-random==1.0.1
numpy==1.14.2
opencv-python==3.4.0.12
Pillow==5.1.0
protobuf==3.5.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
scipy==1.1.0
six==1.11.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
urllib3==1.22
Werkzeug==0.14.1
